+++ backend/src/db/score/score.dto.ts
@@ -17,31 +17,8 @@ export class BaseScoreDto {
   gameMode: string;
 }
 
-export class CreateScoreRequest extends BaseScoreDto {}
-export class CreateScoreInsert extends BaseScoreDto {
+export class InsertScoreDto extends BaseScoreDto {
   @IsString()
   @IsNotEmpty()
   uuid: string;
-}
-
-export class CreateScoreResponse {
-  allCount: number;
-  rank: number;
-  uuid: string;
-}
-
-export class GetScoreRequest {
-  @IsString()
-  @IsNotEmpty()
-  gameMode: string;
-}
-
-export class UpdateScoreNameRequest {
-  @IsString()
-  @IsNotEmpty()
-  uuid: string;
-
-  @IsString()
-  @IsNotEmpty()
-  userName: string;
 }
